requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community
langchain-google-community
openai
pydantic
python-dotenv
requests
google-api-python-client
python-multipart
streamlit
plotly
pandas
